requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
geopandas==0.14.4
sqlalchemy==2.0.23
GeoAlchemy2==0.14.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fiona==1.10.1
werkzeug==3.0.1
